Refactor SDK syntax files to be similar to cats.syntax.all
