+++ casper/src/main/scala/coop/rchain/casper/BlockExecutionTracker.scala
@@ -5,7 +5,7 @@ import cats.effect.concurrent.Ref
 import cats.syntax.all._
 import coop.rchain.blockstorage.dag.BlockDagStorage.DeployId
 import coop.rchain.rholang.interpreter.EvaluateResult
-import coop.rchain.sdk._
+import coop.rchain.sdk.syntax.all._
 
 trait BlockExecutionTracker[F[_]] {
   def execStarted(d: DeployId): F[Unit]
