--- sdk/src/main/scala/coop/rchain/sdk/package.scala
@@ -1,3 +1,0 @@
-package coop.rchain
-
-package object sdk extends ThrowableSyntax
