+++ sdk/src/main/scala/coop/rchain/sdk/syntax/all.scala
@@ -1,0 +1,5 @@
+package coop.rchain.sdk.syntax
+
+import coop.rchain.sdk.primitive._
+
+trait AllSyntax extends ThrowableSyntax
